% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{canvas_api_key}
\alias{canvas_api_key}
\title{Get the Canvas API key from the environment variable}
\usage{
canvas_api_key()
}
\value{
The Canvas API key stored in the CANVAS_API_KEY environment variable.
}
\description{
Get the Canvas API key from the environment variable
}
