% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{canvas_base_url}
\alias{canvas_base_url}
\title{Get the Canvas base URL from the environment variable}
\usage{
canvas_base_url()
}
\value{
The Canvas base URL stored in the CANVAS_BASE_URL environment variable.
}
\description{
Get the Canvas base URL from the environment variable
}
