% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_assignment_group.R
\name{create_assignment_group}
\alias{create_assignment_group}
\title{Create an Assignment Group in Canvas LMS}
\usage{
create_assignment_group(canvas, course_id, group_name, group_position = NULL)
}
\arguments{
\item{canvas}{A list containing the 'api_key' and 'base_url' for authentication.}

\item{course_id}{The ID of the course in which to create the assignment group.}

\item{group_name}{The name of the assignment group.}

\item{group_position}{The position of the assignment group in the course (optional).}
}
\value{
A confirmation message that the assignment group has been created.
}
\description{
Creates a new assignment group in a specific course using the Canvas LMS API.
}
