% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_course_datalake.R
\name{create_course_datalake}
\alias{create_course_datalake}
\title{Create a data lake for a course.}
\usage{
create_course_datalake(canvas, course_id, storage_location)
}
\arguments{
\item{canvas}{An object containing the Canvas API key and base URL, obtained through the \code{canvas_authenticate} function.}

\item{course_id}{The ID of the course for which to create the data lake.}

\item{storage_location}{The path to the storage location where the data files will be saved.}
}
\value{
NULL.
}
\description{
This function retrieves data from various endpoints for a specific course in the Canvas LMS API
and stores the data as JSON files in a specified storage location.
}
\note{
This function retrieves data from various endpoints. Access to certain endpoints may require specific roles.
}
