% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_course_section.R
\name{create_course_section}
\alias{create_course_section}
\title{Create a Course Section in Canvas LMS}
\usage{
create_course_section(
  canvas,
  course_id,
  section_name,
  section_start_date = NULL,
  section_end_date = NULL
)
}
\arguments{
\item{canvas}{A list containing the 'api_key' and 'base_url' for authentication.}

\item{course_id}{The ID of the course in which to create the section.}

\item{section_name}{The name of the section.}

\item{section_start_date}{(Optional) The start date of the section.}

\item{section_end_date}{(Optional) The end date of the section.}
}
\value{
A confirmation message that the section has been created.
}
\description{
Creates a new course section in a specific course using the Canvas LMS API.
}
