% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_group_category.R
\name{create_group_category}
\alias{create_group_category}
\title{Create a Group Category in Canvas LMS}
\usage{
create_group_category(
  canvas,
  course_id,
  category_name,
  allow_self_signup = FALSE
)
}
\arguments{
\item{canvas}{A list containing the 'api_key' and 'base_url' for authentication.}

\item{course_id}{The ID of the course in which to create the group category.}

\item{category_name}{The name of the group category.}

\item{allow_self_signup}{(Optional) Whether to allow self-signup for groups in the category. Defaults to FALSE.}
}
\value{
A confirmation message that the group category has been created.
}
\description{
Creates a new group category in a specific course using the Canvas LMS API.
}
