% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_module.R
\name{create_module}
\alias{create_module}
\title{Create a Module in Canvas LMS}
\usage{
create_module(canvas, course_id, module_name, position = NULL)
}
\arguments{
\item{canvas}{A list containing the 'api_key' and 'base_url' for authentication.}

\item{course_id}{The ID of the course in which to create the module.}

\item{module_name}{The title of the module.}

\item{position}{(Optional) The position of this module in the course}
}
\value{
A confirmation message that the module has been created.
}
\description{
Creates a new module in a specific course using the Canvas LMS API.
}
