% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_module_item.R
\name{create_module_item}
\alias{create_module_item}
\title{Create a Module Item in Canvas LMS}
\usage{
create_module_item(
  canvas,
  course_id,
  module_id,
  item_title,
  item_type = "Page",
  position = NULL,
  page_url = NULL,
  page_id = NULL
)
}
\arguments{
\item{canvas}{A list containing the 'api_key' and 'base_url' for authentication.}

\item{course_id}{The ID of the course in which to create the module item.}

\item{module_id}{The ID of the module in which to create the item.}

\item{item_title}{The title of the new item within the module.}

\item{item_type}{The type of item. Defaults to "Page".}

\item{position}{(Optional) The position of the item within the module. Defaults to 1.}

\item{page_url}{(Optional) The url of the page.}

\item{page_id}{(Optional) The id of the page.}
}
\value{
A confirmation message that the page has been created.
}
\description{
Creates a new item in a specific module in a specifc course using the Canvas LMS API.
}
