% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_page.R
\name{create_page}
\alias{create_page}
\title{Create a Page in Canvas LMS}
\usage{
create_page(canvas, course_id, page_title, page_body, published = FALSE)
}
\arguments{
\item{canvas}{A list containing the 'api_key' and 'base_url' for authentication.}

\item{course_id}{The ID of the course in which to create the page.}

\item{page_title}{The title of the page.}

\item{page_body}{The body/content of the page.}

\item{published}{(Optional) Whether the page should be published. Defaults to FALSE.}
}
\value{
A confirmation message that the page has been created.
}
\description{
Creates a new page in a specific course using the Canvas LMS API.
}
