% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_course_section.R
\name{delete_course_section}
\alias{delete_course_section}
\title{Delete a Course Section in Canvas LMS}
\usage{
delete_course_section(canvas, section_id)
}
\arguments{
\item{canvas}{A list containing the 'api_key' and 'base_url' for authentication.}

\item{section_id}{The ID of the section to delete.}
}
\value{
A confirmation message that the section has been deleted.
}
\description{
Deletes an existing course section using the Canvas LMS API.
}
