% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paginate.R
\name{extract_next_url}
\alias{extract_next_url}
\title{Extracts the 'next' URL from a Link header}
\usage{
extract_next_url(link_header)
}
\arguments{
\item{link_header}{The Link header string from an httr response.}
}
\value{
The URL (character) for the next page, or NULL if not present.
}
\description{
Extracts the 'next' URL from a Link header
}
