% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_account_rubric.R
\name{get_account_rubric}
\alias{get_account_rubric}
\title{Get Single Account Rubric from Canvas LMS API}
\usage{
get_account_rubric(canvas, account_id, rubric_id)
}
\arguments{
\item{canvas}{A list containing the 'api_key' and 'base_url' for authentication.}

\item{account_id}{The ID of the account containing the rubric.}

\item{rubric_id}{The ID of the rubric to retrieve.}
}
\value{
A rubric object retrieved from the Canvas LMS API.
}
\description{
Fetches a single rubric by ID from a specific account in the Canvas LMS API.
}