% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_account_rubrics.R
\name{get_account_rubrics}
\alias{get_account_rubrics}
\title{Get Account Rubrics from Canvas LMS API}
\usage{
get_account_rubrics(canvas, account_id, per_page = 100)
}
\arguments{
\item{canvas}{A list containing the 'api_key' and 'base_url' for authentication.}

\item{account_id}{The ID of the account for which to retrieve the rubrics.}

\item{per_page}{Number of rubrics to retrieve per page. Default is 100.}
}
\value{
A list of rubrics retrieved from the Canvas LMS API.
}
\description{
Fetches a list of rubrics within a specific account from the Canvas LMS API.
}