% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_assignment_groups.R
\name{get_assignment_groups}
\alias{get_assignment_groups}
\title{Retrieves the assignment groups within a course.}
\usage{
get_assignment_groups(canvas, course_id, per_page = 100)
}
\arguments{
\item{canvas}{An object containing the Canvas API key and base URL, obtained through the \code{canvas_authenticate} function.}

\item{course_id}{The ID of the course for which to fetch the assignment groups.}

\item{per_page}{The number of entries to show per page.}
}
\value{
A list of assignment groups within the specified course.
}
\description{
This function retrieves the assignment groups within a specific course in the Canvas LMS API.
}
