% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_assignment_submissions.R
\name{get_assignment_submissions}
\alias{get_assignment_submissions}
\title{Retrieves assignment submissions.}
\usage{
get_assignment_submissions(canvas, course_id, assignment_id, per_page = 100)
}
\arguments{
\item{canvas}{An object containing the Canvas API key and base URL, obtained through the \code{canvas_authenticate} function.}

\item{course_id}{The ID of the course for which to fetch submissions.}

\item{assignment_id}{The ID of the assignment for which to fetch the submissions.}

\item{per_page}{(Optional) The number of submissions to retrieve per page of results (default is 100).}
}
\value{
A data frame containing the submission data.
}
\description{
This function retrieves the submissions to a specific assignment in a specific course in the
Canvas LMS API.
}
