% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_assignments.R
\name{get_assignments}
\alias{get_assignments}
\title{Get Assignments from Canvas LMS API}
\usage{
get_assignments(canvas, course_id)
}
\arguments{
\item{canvas}{A list containing the 'api_key' and 'base_url' for authentication.}

\item{course_id}{The ID of the course for which to retrieve the assignments.}
}
\value{
A list of assignments retrieved from the Canvas LMS API.
}
\description{
Fetches a list of assignments within a specific course from the Canvas LMS API.
}
