% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_conversations.R
\name{get_conversations}
\alias{get_conversations}
\title{Retrieves conversations.}
\usage{
get_conversations(canvas)
}
\arguments{
\item{canvas}{An object containing the Canvas API key and base URL, obtained through the
\code{canvas_authenticate} function.}
}
\value{
Returns a data frame containing the details of conversations for the current user, most
recent ones first.
}
\description{
This function retrieves all details concerning conversations of the user.
}
\seealso{
\code{\link[=get_single_conversation]{get_single_conversation()}} and \code{\link[=create_conversation]{create_conversation()}}.
}
