% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_course_announcements.R
\name{get_course_announcements}
\alias{get_course_announcements}
\title{Retrieves course announcements.}
\usage{
get_course_announcements(canvas, course_id, per_page = 100)
}
\arguments{
\item{canvas}{An object containing the Canvas API key and base URL, obtained through the \code{canvas_authenticate} function.}

\item{course_id}{The ID of the course for which to retrieve announcements.}

\item{per_page}{(Optional) The number of announcements to retrieve per page of results (default is 100).}
}
\value{
A data frame of course announcements.
}
\description{
This function retrieves a list of announcements for a specific course in the Canvas LMS API.
}
