% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_course_enrollments.R
\name{get_course_enrollments}
\alias{get_course_enrollments}
\title{Retrieves the course enrollments for a course.}
\usage{
get_course_enrollments(canvas, course_id, per_page = 100)
}
\arguments{
\item{canvas}{An object containing the Canvas API key and base URL, obtained through the \code{canvas_authenticate} function.}

\item{course_id}{The ID of the course for which to fetch the enrollments.}

\item{per_page}{(Optional) The number of enrollments to retrieve per page of results (default is 100).}
}
\value{
A data frame of course enrollments for the specified course.
}
\description{
This function retrieves the enrollments of students and other roles in a specific course in the Canvas LMS API.
}
