% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_course_files.R
\name{get_course_files}
\alias{get_course_files}
\title{Retrieves a list of files within a course.}
\usage{
get_course_files(canvas, course_id, per_page = 100)
}
\arguments{
\item{canvas}{An object containing the Canvas API key and base URL, obtained through the \code{canvas_authenticate} function.}

\item{course_id}{The ID of the course for which to fetch the files.}

\item{per_page}{(Optional) The number of files to retrieve per page of results (default is 100).}
}
\value{
A data frame of files within the specified course.
}
\description{
This function retrieves a list of files within a specific course in the Canvas LMS API.
}
