% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_course_groups.R
\name{get_course_groups}
\alias{get_course_groups}
\title{Retrieves the list of groups in a course.}
\usage{
get_course_groups(canvas, course_id, per_page = 100)
}
\arguments{
\item{canvas}{A list containing the 'api_key' and 'base_url' for authentication.}

\item{course_id}{The ID of the course for which to fetch the groups.}

\item{per_page}{(Optional) The number of groups to retrieve per page of results (default is 100).}
}
\value{
A data frame of groups in the specified course.
}
\description{
This function retrieves the list of groups in a specific course in the Canvas LMS API.
}
