% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_course_media_objects.R
\name{get_course_media_objects}
\alias{get_course_media_objects}
\title{Retrieves the media objects in a course.}
\usage{
get_course_media_objects(canvas, course_id, per_page = 100)
}
\arguments{
\item{canvas}{An object containing the Canvas API key and base URL, obtained through the \code{canvas_authenticate} function.}

\item{course_id}{The ID of the course for which to fetch the media objects.}

\item{per_page}{The number of entries to show per page.}
}
\value{
A data frame containing the media objects in the specified course.
}
\description{
This function retrieves the media objects associated with a specific course in the Canvas LMS API.
}
