% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_course_pages.R
\name{get_course_pages}
\alias{get_course_pages}
\title{Retrieves the pages within a course.}
\usage{
get_course_pages(canvas, course_id, per_page = 100)
}
\arguments{
\item{canvas}{An object containing the Canvas API key and base URL, obtained through the \code{canvas_authenticate} function.}

\item{course_id}{The ID of the course for which to fetch the pages.}

\item{per_page}{(Optional) The number of pages to retrieve per page of results (default is 50).}
}
\value{
A list of pages within the specified course.
}
\description{
This function retrieves the pages within a specific course in the Canvas LMS API.
}
