% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_course_participation.R
\name{get_course_participation}
\alias{get_course_participation}
\title{Get course-level participation data from Canvas LMS API}
\usage{
get_course_participation(canvas, course_id)
}
\arguments{
\item{canvas}{A list containing the 'api_key' and 'base_url' for authentication.}

\item{course_id}{The ID of the course for which to retrieve the participation data.}
}
\value{
The participation data for the specified course retrieved from the Canvas LMS API.
}
\description{
Fetches the participation data for a specific course from the Canvas LMS API.
}
