% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_course_quizzes.R
\name{get_course_quizzes}
\alias{get_course_quizzes}
\title{Retrieves course quizzes.}
\usage{
get_course_quizzes(canvas, course_id, per_page = 100)
}
\arguments{
\item{canvas}{An object containing the Canvas API key and base URL, obtained through the \code{canvas_authenticate} function.}

\item{course_id}{The ID of the course for which to retrieve quizzes.}

\item{per_page}{(Optional) The number of quizzes to retrieve per page of results (default is 100).}
}
\value{
A data frame of course quizzes.
}
\description{
This function retrieves a list of quizzes for a specific course in the Canvas LMS API.
}
