% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_course_root_folder.R
\name{get_course_root_folder}
\alias{get_course_root_folder}
\title{Retrieves root folder of a course.}
\usage{
get_course_root_folder(canvas, course_id)
}
\arguments{
\item{canvas}{An object containing the Canvas API key and base URL, obtained through the
\code{canvas_authenticate} function.}

\item{course_id}{The ID of the course for which to fetch associated folders.}
}
\value{
Returns a data frame containing the details on the root folder associated with the
specified course.
}
\description{
This function retrieves the root folder associated with a course.
}
\seealso{
\code{\link[=get_course_folders]{get_course_folders()}} and \code{\link[=get_user_folders]{get_user_folders()}}
}
