% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_course_sections.R
\name{get_course_sections}
\alias{get_course_sections}
\title{Retrieves course sections.}
\usage{
get_course_sections(canvas, course_id, per_page = 100)
}
\arguments{
\item{canvas}{An object containing the Canvas API key and base URL, obtained through the \code{canvas_authenticate} function.}

\item{course_id}{The ID of the course for which to retrieve sections.}

\item{per_page}{(Optional) The number of sections to retrieve per page of results (default is 100).}
}
\value{
A data frame of course sections.
}
\description{
This function retrieves a list of sections for a specific course in the Canvas LMS API.
}
