% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_course_students.R
\name{get_course_students}
\alias{get_course_students}
\title{Retrieves the list of students in a course.}
\usage{
get_course_students(canvas, course_id, per_page = 100)
}
\arguments{
\item{canvas}{An object containing the Canvas API key and base URL, obtained through the \code{canvas_authenticate} function.}

\item{course_id}{The ID of the course for which to fetch the students.}

\item{per_page}{(Optional) The number of students to retrieve per page of results (default is 100).}
}
\value{
A data frame of students enrolled in the specified course.
}
\description{
This function retrieves the list of students enrolled in a specific course in the Canvas LMS API.
}
