% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_courses.R
\name{get_courses}
\alias{get_courses}
\title{Get Courses from Canvas LMS API}
\usage{
get_courses(canvas, per_page = 100)
}
\arguments{
\item{canvas}{A list containing the 'api_key' and 'base_url' for authentication.}

\item{per_page}{Number of courses to retrieve per page. Default is 100.}
}
\value{
A list of courses retrieved from the Canvas LMS API.
}
\description{
Retrieves a list of courses from the Canvas LMS API.
}
