% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_department_statistics.R
\name{get_department_statistics}
\alias{get_department_statistics}
\title{Get department-level statistics from the Canvas LMS API}
\usage{
get_department_statistics(canvas, account_id, type = "current", term_id = NULL)
}
\arguments{
\item{canvas}{A list containing the 'api_key' and 'base_url' for authentication.}

\item{account_id}{The ID of the account for which to retrieve the statistics.}

\item{type}{The type of courses to include in the data. Can be 'current', 'completed', or 'term'.}

\item{term_id}{The ID of the term for which to retrieve the statistics. Only used when type is 'terms/<term_id>'.}
}
\value{
A list of department-level statistics retrieved from the Canvas LMS API.
}
\description{
Retrieves department-level statistics for a specific account and term from the Canvas LMS API.
}
