% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_discussions.R
\name{get_discussions}
\alias{get_discussions}
\title{Retrieves the discussion topics within a course.}
\usage{
get_discussions(canvas, course_id, per_page = 100)
}
\arguments{
\item{canvas}{An object containing the Canvas API key and base URL, obtained through the \code{canvas_authenticate} function.}

\item{course_id}{The ID of the course for which to fetch the discussion topics.}

\item{per_page}{The number of entries to show}
}
\value{
A list of discussion topics within the specified course.
}
\description{
This function retrieves the discussion topics within a specific course in the Canvas LMS API.
}
\examples{
# Authenticate with the Canvas LMS API
canvas <- canvas_authenticate(api_key, base_url)

# Fetch the discussion topics of a specific course
course_id <- 12345
discussions <- get_discussions(canvas, course_id)
}
