% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_favorite_courses.R
\name{get_favorite_courses}
\alias{get_favorite_courses}
\title{Get Favorite Courses in Canvas LMS}
\usage{
get_favorite_courses(canvas, exclude_blueprint_courses = NULL)
}
\arguments{
\item{canvas}{A list containing the 'api_key' and 'base_url' for authentication.}

\item{exclude_blueprint_courses}{When set, only return courses that are not configured as blueprint courses (optional).}
}
\value{
The dataframe of favorite courses.
}
\description{
Retrieves the data of favorite courses for the authenticated user using the Canvas LMS API.
}
