% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_folder_files.R
\name{get_folder_files}
\alias{get_folder_files}
\title{Retrieves files in a specific folder.}
\usage{
get_folder_files(canvas, folder_id)
}
\arguments{
\item{canvas}{An object containing the Canvas API key and base URL, obtained through the
\code{canvas_authenticate} function.}

\item{folder_id}{The ID of the folder for which to fetch associated files.}
}
\value{
Returns a data frame containing the details on the files stored in the specified folder.
}
\description{
This function lists all files available in a specific folder.
}
