% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_group_categories.R
\name{get_group_categories}
\alias{get_group_categories}
\title{Get group categories for a context}
\usage{
get_group_categories(canvas, course_id)
}
\arguments{
\item{canvas}{A list containing the 'api_key' and 'base_url' for authentication.}

\item{course_id}{The ID of the course for which to retrieve the group categories.}
}
\value{
A data frame of group categories in the specified context.
}
\description{
This function retrieves the group categories for a specific context (e.g., course) in the Canvas LMS API.
}
