% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_group_info.R
\name{get_group_info}
\alias{get_group_info}
\title{Get information about a single group}
\usage{
get_group_info(canvas, group_id)
}
\arguments{
\item{canvas}{A list containing the 'api_key' and 'base_url' for authentication.}

\item{group_id}{The ID of the group for which to retrieve the information.}
}
\value{
A list containing the information about the specified group.
}
\description{
This function retrieves information about a specific group in the Canvas LMS API.
}
