% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_group_memberships.R
\name{get_group_memberships}
\alias{get_group_memberships}
\title{Get group memberships}
\usage{
get_group_memberships(canvas, group_id)
}
\arguments{
\item{canvas}{A list containing the 'api_key' and 'base_url' for authentication.}

\item{group_id}{The ID of the group for which to retrieve the memberships.}
}
\value{
A data frame of memberships in the specified group.
}
\description{
This function retrieves the memberships for a specific group in the Canvas LMS API.
}
