% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_group_users.R
\name{get_group_users}
\alias{get_group_users}
\title{Get users in a group}
\usage{
get_group_users(canvas, group_id)
}
\arguments{
\item{canvas}{A list containing the 'api_key' and 'base_url' for authentication.}

\item{group_id}{The ID of the group for which to retrieve the users.}
}
\value{
A data frame of users in the specified group.
}
\description{
This function retrieves the users in a specific group in the Canvas LMS API.
}
