% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_modules.R
\name{get_modules}
\alias{get_modules}
\title{Retrieves the modules within a course.}
\usage{
get_modules(canvas, course_id, per_page = 100)
}
\arguments{
\item{canvas}{An object containing the Canvas API key and base URL, obtained through the \code{canvas_authenticate} function.}

\item{course_id}{The ID of the course for which to fetch the modules.}

\item{per_page}{The number of entries to show}
}
\value{
A list of modules within the specified course.
}
\description{
This function retrieves the modules within a specific course in the Canvas LMS API.
}
