% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_quiz_submissions.R
\name{get_quiz_submissions}
\alias{get_quiz_submissions}
\title{Retrieves quiz submissions.}
\usage{
get_quiz_submissions(canvas, course_id, quiz_id)
}
\arguments{
\item{canvas}{An object containing the Canvas API key and base URL, obtained through the
\code{canvas_authenticate} function.}

\item{course_id}{The ID of the course for which to fetch submissions.}

\item{quiz_id}{The ID of the quiz for which to fetch the submissions.}
}
\value{
A data frame containing the submission data.
}
\description{
This function retrieves the submissions to a specific quiz in a specific course in the Canvas
LMS API.
}
\seealso{
\code{\link[=get_assignment_submissions]{get_assignment_submissions()}}
}
