% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_roles.R
\name{get_roles}
\alias{get_roles}
\title{Retrieve Roles for a Canvas Account}
\usage{
get_roles(canvas, account_id, per_page = 100)
}
\arguments{
\item{canvas}{A list containing the base URL and API key for the Canvas LMS instance.}

\item{account_id}{The ID of the account for which you want to retrieve the roles.}

\item{per_page}{The number of roles to retrieve per page (default is 100).}
}
\value{
A data frame containing the roles available to the specified account.
}
\description{
This function retrieves a paginated list of the roles available to a specific account in the Canvas LMS system.
}
