% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rubric.R
\name{get_rubric}
\alias{get_rubric}
\title{Get Single Rubric from Canvas LMS API}
\usage{
get_rubric(canvas, course_id, rubric_id)
}
\arguments{
\item{canvas}{A list containing the 'api_key' and 'base_url' for authentication.}

\item{course_id}{The ID of the course containing the rubric.}

\item{rubric_id}{The ID of the rubric to retrieve.}
}
\value{
A rubric object retrieved from the Canvas LMS API.
}
\description{
Fetches a single rubric by ID from a specific course in the Canvas LMS API.
}