% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rubrics.R
\name{get_rubrics}
\alias{get_rubrics}
\title{Get Rubrics from Canvas LMS API}
\usage{
get_rubrics(canvas, course_id, per_page = 100)
}
\arguments{
\item{canvas}{A list containing the 'api_key' and 'base_url' for authentication.}

\item{course_id}{The ID of the course for which to retrieve the rubrics.}

\item{per_page}{Number of rubrics to retrieve per page. Default is 100.}
}
\value{
A list of rubrics retrieved from the Canvas LMS API.
}
\description{
Fetches a list of rubrics within a specific course from the Canvas LMS API.
}