% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_section_information.R
\name{get_section_information}
\alias{get_section_information}
\title{Get Section Information in Canvas LMS}
\usage{
get_section_information(canvas, course_id, section_id)
}
\arguments{
\item{canvas}{A list containing the 'api_key' and 'base_url' for authentication.}

\item{course_id}{The ID of the course.}

\item{section_id}{The ID of the section.}
}
\value{
The information about the section.
}
\description{
Retrieves information about a specific course section using the Canvas LMS API.
}
