% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_single_conversation.R
\name{get_single_conversation}
\alias{get_single_conversation}
\title{Retrieves a single conversation.}
\usage{
get_single_conversation(canvas, conversation_id)
}
\arguments{
\item{canvas}{An object containing the Canvas API key and base URL, obtained through the
\code{canvas_authenticate} function.}

\item{conversation_id}{The id of the specific conversation.}
}
\value{
Returns information for a specific conversation for the current user. Response includes
all fields that are present in the list/index action as well as messages and extended
participant information.
}
\description{
This function retrieves all details concerning a specific conversation.
}
\seealso{
\code{\link[=get_conversations]{get_conversations()}} and \code{\link[=create_conversation]{create_conversation()}}.
}
