% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_student_summaries.R
\name{get_student_summaries}
\alias{get_student_summaries}
\title{Get student summaries for a course from Canvas LMS API}
\usage{
get_student_summaries(canvas, course_id, per_page = 100)
}
\arguments{
\item{canvas}{A list containing the 'api_key' and 'base_url' for authentication.}

\item{course_id}{The ID of the course for which to retrieve the student summaries.}

\item{per_page}{Number of student summaries to retrieve per page. Default is 100.}
}
\value{
A list of student summaries retrieved from the Canvas LMS API.
}
\description{
Retrieves the student summaries for a specific course from the Canvas LMS API.
}
