% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_course_assignment_data.R
\name{get_user_course_assignment_data}
\alias{get_user_course_assignment_data}
\title{Get user-in-a-course-level assignment data from the Canvas LMS API}
\usage{
get_user_course_assignment_data(canvas, course_id, student_id)
}
\arguments{
\item{canvas}{A list containing the 'api_key' and 'base_url' for authentication.}

\item{course_id}{The ID of the course for which to retrieve the assignment data.}

\item{student_id}{The ID of the student for which to retrieve the assignment data.}
}
\value{
A list of user-in-a-course-level assignment data retrieved from the Canvas LMS API.
}
\description{
Retrieves user-in-a-course-level assignment data for a specific course and student from the Canvas LMS API.
}
