% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_folders.R
\name{get_user_folders}
\alias{get_user_folders}
\title{Retrieves folders of the current user.}
\usage{
get_user_folders(canvas)
}
\arguments{
\item{canvas}{An object containing the Canvas API key and base URL, obtained through the
\code{canvas_authenticate} function.}
}
\value{
Returns a data frame containing the details on the folders available to the user
accessing the API.
}
\description{
This function retrieves all folders associated with the user accessing the API.
}
\details{
Returns a list of all folders 'owned by' the current user. Note that since \code{self} replaces
\code{user_id}, the id of the user accessing the API does not need to be specified. The
highest level folder name is \verb{my folder}.
}
\seealso{
\code{\link[=get_course_root_folder]{get_course_root_folder()}} and \code{\link[=get_course_folders]{get_course_folders()}}
}
