% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_all_enrollment_terms.R
\name{list_all_enrollment_terms}
\alias{list_all_enrollment_terms}
\title{List All Enrollment Terms}
\usage{
list_all_enrollment_terms(canvas, account_id, per_page = 100)
}
\arguments{
\item{canvas}{A list containing Canvas API configuration:
\itemize{
\item base_url: The base URL of your Canvas instance
\item api_key: Your Canvas API key
}}

\item{account_id}{The ID of the account for which to retrieve enrollment terms}

\item{per_page}{The number of terms to retrieve per page (default is 100)}
}
\value{
A dataframe containing all enrollment terms details
}
\description{
This function retrieves a paginated list of all enrollment terms for a specified account in Canvas LMS.
}
