% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paginate.R
\name{paginate}
\alias{paginate}
\title{Helper to paginate Canvas API GET requests}
\usage{
paginate(initial_response, access_token, showProgress = TRUE)
}
\description{
Follows 'next' links in the Link header to retrieve all pages.
Returns a list of all responses.
}
