% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_new_discussion.R
\name{post_new_discussion}
\alias{post_new_discussion}
\title{Post a New Discussion in Canvas LMS}
\usage{
post_new_discussion(
  canvas,
  course_id,
  discussion_title,
  discussion_message,
  discussion_is_announcement = FALSE
)
}
\arguments{
\item{canvas}{A list containing the 'api_key' and 'base_url' for authentication.}

\item{course_id}{The ID of the course in which to create the discussion.}

\item{discussion_title}{The title of the discussion topic.}

\item{discussion_message}{The initial message content of the discussion topic.}

\item{discussion_is_announcement}{(Optional) Whether the discussion should be an announcement. Defaults to FALSE.}
}
\value{
A confirmation message that the discussion has been created.
}
\description{
Creates a new discussion topic in a specific course using the Canvas LMS API.
}
