% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_progress.R
\name{query_progress}
\alias{query_progress}
\title{Queries progress.}
\usage{
query_progress(canvas, progress_url, attempts = 10)
}
\arguments{
\item{canvas}{An object containing the Canvas API key and base URL, obtained through the
\code{canvas_authenticate} function.}

\item{progress_url}{The url where a progress update can be retrieved.}

\item{attempts}{The number of times it is allowed to check progress.}
}
\value{
a message indicating completion and status information about the asynchronous job.
}
\description{
This function queries the progress of asynchronous API operations.
}
\details{
It performs 10 attempts with a pause of 6 seconds between them.
}
