% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_file.R
\name{upload_file}
\alias{upload_file}
\title{Upload a file in Canvas LMS.}
\usage{
upload_file(canvas, course_id, file_name)
}
\arguments{
\item{canvas}{An object containing the Canvas API key and base URL, obtained through the \code{canvas_authenticate} function.}

\item{course_id}{The ID of the course where the file will be uploaded.}

\item{file_name}{The file to be uploaded.}
}
\value{
A confirmation message indicating that the file has been uploaded.
}
\description{
This function uploads a file to the files folder of a specific course.
}
