% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_qti_file_with_migration.R
\name{upload_qti_file_with_migration}
\alias{upload_qti_file_with_migration}
\title{Upload QTI File with Content Migration}
\usage{
upload_qti_file_with_migration(
  canvas,
  course_id,
  pre_attachment_name,
  pre_attachment_size,
  file_path,
  settings = list()
)
}
\arguments{
\item{canvas}{A list containing the 'api_key' and 'base_url' for authentication.}

\item{course_id}{The ID of the course where the QTI file will be uploaded.}

\item{pre_attachment_name}{The name of the file being uploaded. This is required for the file upload process.}

\item{pre_attachment_size}{The size of the file being uploaded in bytes. This is required for the file upload process.}

\item{file_path}{The local file path to the QTI file that needs to be uploaded.}

\item{settings}{A list of additional settings for the content migration. This can include options like question bank name or overwrite settings.}
}
\value{
A confirmation message indicating that the QTI file has been uploaded and the content migration has been completed.
}
\description{
This function uploads a QTI file to a specified course in Canvas LMS and initiates a content migration.
It handles the process of creating the migration, uploading the file, and checking the migration status.
}
