% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_module_items.R
\name{get_module_items}
\alias{get_module_items}
\title{Retrieves the items within a specific module.}
\usage{
get_module_items(canvas, course_id, module_id)
}
\arguments{
\item{canvas}{An object containing the Canvas API key and base URL, obtained through the \code{canvas_authenticate} function.}

\item{course_id}{The ID of the course containing the module.}

\item{module_id}{The ID of the module for which to fetch the items.}
}
\value{
A list of items within the specified module.
}
\description{
This function retrieves the items within a specific module of a course in the Canvas LMS API.
}
