% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paginate.R
\name{paginate}
\alias{paginate}
\title{Helper to paginate Canvas API GET requests}
\usage{
paginate(initial_response, access_token, showProgress = TRUE)
}
\arguments{
\item{initial_response}{The initial httr response object from the first GET request.}

\item{access_token}{The Canvas API access token (string).}

\item{showProgress}{Logical; whether to print progress as pages are fetched. Default is TRUE.}
}
\value{
A list of httr response objects, one for each page.
}
\description{
Follows 'next' links in the Link header to retrieve all pages.
Returns a list of all responses.
}
